Introduce new package pre-commit-excludes
