+++ packages/pre_commit_excludes/LICENSE
@@ -1,0 +1,1 @@
+../../LICENSE
