+++ packages/pre_commit_excludes/README.md
@@ -1,0 +1,7 @@
+# Pre-Commit-Excludes
+
+<!-- rumdl-disable MD013 -->
+[![PyPI](https://img.shields.io/pypi/v/pre-commit-excludes)](https://pypi.org/project/pre-commit-excludes/)
+[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/pre-commit-excludes)](https://pypi.org/project/pre-commit-excludes/)
+[![PyPI - License](https://img.shields.io/pypi/l/pre-commit-excludes)](https://pypi.org/project/pre-commit-excludes/)
+<!-- rumdl-enable MD013 -->
