+++ packages/pre_commit_excludes/pre_commit_excludes/__init__.py
@@ -1,0 +1,1 @@
+"""Tools to maintain pre-commit exclude lists."""
