+++ packages/pre_commit_excludes/pre_commit_excludes/hook_utils.py
@@ -1,0 +1,96 @@
+from __future__ import annotations
+
+import itertools
+from collections import Counter
+from pathlib import Path
+from typing import Any
+
+from ruamel.yaml import YAML
+
+
+class Hook:
+    """Represent a pre-commit hook with its excluded paths."""
+
+    def __init__(self, id: str, exclude_paths: list[Path]) -> None:
+        self.__id = id
+        self.__exclude_paths = exclude_paths
+
+    @classmethod
+    def from_hook_config(cls: type[Hook], root_directory: Path, hook_config: dict[str, str]) -> Hook:
+        excluded_paths_list = [
+            root_directory / Path(exclude) for exclude in extract_literal_exclude_paths(hook_config["exclude"])
+        ]
+
+        return cls(hook_config["id"], excluded_paths_list)
+
+    @property
+    def id(self) -> str:
+        return self.__id
+
+    @property
+    def exclude_paths(self) -> list[Path]:
+        return self.__exclude_paths
+
+    def find_duplicates(self) -> list[Path]:
+        counter = Counter(self.exclude_paths)
+
+        return [path for path in counter if counter[path] > 1]
+
+    def find_non_existing_paths(self) -> list[Path]:
+        return [path for path in self.exclude_paths if not path.resolve().exists()]
+
+    def has_duplicates(self) -> bool:
+        return bool(self.find_duplicates())
+
+    def has_non_existing_paths(self) -> bool:
+        return bool(self.find_non_existing_paths())
+
+    def count_excluded_files(self) -> int:
+        existing_paths = [path for path in self.exclude_paths if path.exists()]
+        total_file_count = sum(1 for path in existing_paths if path.is_file())
+        excluded_dirs = [path for path in existing_paths if path.is_dir()]
+        total_dir_count = sum(
+            sum(1 for file in excluded_dir.rglob("*") if file.is_file()) for excluded_dir in excluded_dirs
+        )
+        return total_file_count + total_dir_count
+
+
+def is_regex_pattern(exclude: str) -> bool:
+    return any(regex_key in exclude for regex_key in ["*", "$", "^"])
+
+
+def extract_literal_exclude_paths(exclude_regex: str) -> list[str]:
+    exclude_list = (
+        _remove_verbose_regex_comments(exclude_regex)
+        .replace("\n", "")
+        .replace(" ", "")
+        .replace("(?x)^(", "")
+        .replace("^", "")
+        .replace(r"\.", ".")
+        .replace(")", "")
+        .split("|")
+    )
+
+    return [exclude for exclude in exclude_list if not is_regex_pattern(exclude)]
+
+
+def _remove_verbose_regex_comments(exclude: str) -> str:
+    return "\n".join(line.split("#", 1)[0] for line in exclude.splitlines())
+
+
+def has_excludes(hook_config: dict[str, str]) -> bool:
+    return bool(hook_config.get("exclude")) and hook_config.get("exclude") != "^$"
+
+
+def load_config(config_file: Path) -> dict[str, Any]:
+    yaml = YAML(typ="safe")
+    config = yaml.load(config_file.read_text(encoding="utf-8"))
+
+    return config if isinstance(config, dict) else {}
+
+
+def load_hooks(root_directory: Path, config_file: Path) -> list[Hook]:
+    config = load_config(config_file)
+    hook_configs = itertools.chain(*[repo["hooks"] for repo in config["repos"]])
+
+    return [Hook.from_hook_config(root_directory, hook) for hook in hook_configs if has_excludes(hook)]
