+++ tests/pre_commit_excludes/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for packages/pre_commit_excludes."""
