+++ tests/pre_commit_excludes/test_hook_utils.py
@@ -1,0 +1,335 @@
+from __future__ import annotations
+
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+import pytest
+from pre_commit_excludes.hook_utils import (
+    Hook,
+    extract_literal_exclude_paths,
+    has_excludes,
+    is_regex_pattern,
+    load_hooks,
+)
+from ruamel.yaml import YAML
+
+if TYPE_CHECKING:
+    from pyfakefs.fake_filesystem import FakeFilesystem
+
+
+@pytest.mark.parametrize("pattern", [".*\\/conanfile.py", "\\.lock$", "^samples/ros1/src/"])
+def test_is_regex_pattern_for_regex_should_be_true(pattern: str) -> None:
+    assert is_regex_pattern(pattern)
+
+
+def test_is_regex_pattern_for_no_regex_should_be_false() -> None:
+    assert is_regex_pattern("packages/thirdparty/") is False
+
+
+def test_extract_literal_exclude_paths_should_ignore_regex_patterns_and_comments() -> None:
+    assert extract_literal_exclude_paths(
+        r"""(?x)^(
+  # keep-sorted start ignore_prefixes=*
+  BUILD.bazel|
+  bar/.*\.png|
+  # keep-sorted end *
+  foo.txt
+)""",
+    ) == ["BUILD.bazel", "foo.txt"]
+
+
+def test_extract_literal_exclude_paths_for_regex_only_exclude_should_return_empty_list() -> None:
+    assert extract_literal_exclude_paths(r"(?x)^(bar/.*\.png|.*\.lock$)") == []
+
+
+def test_extract_literal_exclude_paths_for_escaped_dot_should_unescape() -> None:
+    assert extract_literal_exclude_paths(r"(?x)^(foo/bar\.py)") == ["foo/bar.py"]
+
+
+def test_extract_literal_exclude_paths_for_single_literal_should_return_path() -> None:
+    assert extract_literal_exclude_paths("packages/thirdparty/") == ["packages/thirdparty/"]
+
+
+def test_extract_literal_exclude_paths_for_multiple_literals_should_return_paths() -> None:
+    assert extract_literal_exclude_paths("(?x)^(python/aws_auth|packages/thirdparty/)") == [
+        "python/aws_auth",
+        "packages/thirdparty/",
+    ]
+
+
+def test_from_hook_config_for_single_path(fs: FakeFilesystem) -> None:
+    root_directory = Path("Test_directory/")
+    fs.create_dir(root_directory)
+    hook_instance = Hook.from_hook_config(root_directory, {"id": "buildifier", "exclude": "packages/thirdparty/"})
+
+    assert has_excludes({"id": "check-snake-case", "exclude": "packages/thirdparty"})
+    assert hook_instance.id == "buildifier"
+    assert hook_instance.exclude_paths == [root_directory / Path("packages/thirdparty/")]
+
+
+def test_from_hook_config_for_multiple_paths(fs: FakeFilesystem) -> None:
+    root_directory = Path("Test_directory/")
+    fs.create_dir(root_directory)
+    hook_instance = Hook.from_hook_config(
+        root_directory,
+        {"id": "buildifier", "exclude": "(?x)^(\n  packages/thirdparty/|\n  python/aws_auth/\n)\n"},
+    )
+
+    assert hook_instance.id == "buildifier"
+    assert hook_instance.exclude_paths == [
+        root_directory / Path("packages/thirdparty/"),
+        root_directory / Path("python/aws_auth/"),
+    ]
+
+
+def test_from_hook_config_for_keep_sorted_comments_should_ignore_comments(fs: FakeFilesystem) -> None:
+    root_directory = Path("Test_directory/")
+    fs.create_dir(root_directory)
+    hook_instance = Hook.from_hook_config(
+        root_directory,
+        {
+            "id": "buildifier",
+            "exclude": r"""(?x)^(
+  # keep-sorted start ignore_prefixes=*
+  BUILD.bazel|
+  bar/.*\.png|
+  # keep-sorted end *
+  foo.txt
+)""",
+        },
+    )
+
+    assert hook_instance.id == "buildifier"
+    assert hook_instance.exclude_paths == [
+        root_directory / Path("BUILD.bazel"),
+        root_directory / Path("foo.txt"),
+    ]
+
+
+def test_find_duplicates_for_duplicate_paths_should_return_duplicate_paths(fs: FakeFilesystem) -> None:
+    fs.create_file(Path("Repo/existing_path1"))
+    fs.create_file(Path("Repo/existing_path2"))
+    hook_instance = Hook(
+        "test_id",
+        [Path("existing_path1"), Path("existing_path2"), Path("existing_path1"), Path("existing_path2")],
+    )
+
+    assert hook_instance.has_duplicates()
+    assert hook_instance.find_duplicates() == [Path("existing_path1"), Path("existing_path2")]
+
+
+def test_find_duplicates_for_no_duplicate_paths_should_return_empty_list(fs: FakeFilesystem) -> None:
+    fs.create_file(Path("Repo/existing_path1"))
+    fs.create_file(Path("Repo/existing_path2"))
+    hook_instance = Hook("test_id", [Path("existing_path1"), Path("existing_path2")])
+
+    assert not hook_instance.has_duplicates()
+    assert hook_instance.find_duplicates() == []
+
+
+def test_find_non_existing_paths_for_non_existing_files_should_return_non_existing_files(fs: FakeFilesystem) -> None:
+    fs.create_file(Path("Repo/existing_path"))
+    hook_instance = Hook(
+        "test_id",
+        [Path("Repo/existing_path"), Path("Repo/non_existing_path1"), Path("Repo/non_existing_path2")],
+    )
+
+    assert hook_instance.has_non_existing_paths()
+    assert hook_instance.find_non_existing_paths() == [
+        Path("Repo/non_existing_path1"),
+        Path("Repo/non_existing_path2"),
+    ]
+
+
+def test_find_non_existing_paths_for_existing_files_should_return_empty_list(fs: FakeFilesystem) -> None:
+    fs.create_file(Path("Repo/existing_path1"))
+    fs.create_file(Path("Repo/existing_path2"))
+    hook_instance = Hook("test_id", [Path("Repo/existing_path1"), Path("Repo/existing_path2")])
+
+    assert not hook_instance.has_non_existing_paths()
+    assert hook_instance.find_non_existing_paths() == []
+
+
+def test_has_excludes_for_existing_excludes_should_return_true() -> None:
+    assert has_excludes({"id": "check-snake-case", "exclude": "packages/thirdparty"})
+
+
+def test_has_excludes_for_non_existing_excludes_should_return_false() -> None:
+    assert not has_excludes({"id": "check-snake-case"})
+
+
+def _create_yaml_content(yaml_file: Path, exclude: str, extra_attributes: dict | None = None) -> YAML:
+    if extra_attributes is None:
+        extra_attributes = {}
+
+    yaml = YAML()
+    return yaml.dump(
+        {
+            "repos": [
+                {
+                    "repo": "meta",
+                    "hooks": [
+                        {"id": "check-hooks-apply"},
+                    ],
+                },
+                {
+                    "repo": "local",
+                    "hooks": [
+                        {
+                            "id": "check-snake-case",
+                            "name": "check snake case",
+                            "entry": "python3 foo.py",
+                            "language": "python",
+                            **({"exclude": exclude} if exclude else {}),
+                            **extra_attributes,
+                        },
+                    ],
+                },
+            ],
+        },
+        yaml_file,
+    )
+
+
+def test_load_hooks_for_no_exclude_file(fs: FakeFilesystem) -> None:
+    root_directory = Path("Test_directory/")
+    fs.create_dir(root_directory)
+    config_file = root_directory / Path(".pre-commit-config.yaml")
+    fs.create_file(config_file)
+    _create_yaml_content(config_file, "")
+
+    assert len(load_hooks(root_directory, config_file)) == 0
+
+
+def test_load_hooks_for_exclude_files(fs: FakeFilesystem) -> None:
+    root_directory = Path("Test_directory/")
+    fs.create_dir(root_directory)
+    config_file = root_directory / Path(".pre-commit-config.yaml")
+    fs.create_file(config_file)
+    _create_yaml_content(config_file, "(?x)^(python/aws_auth|packages/thirdparty/)")
+
+    result = load_hooks(root_directory, config_file)
+
+    assert len(result) == 1
+    assert result[0].id == "check-snake-case"
+
+
+def test_load_hooks_for_pass_filenames_42_should_parse_config(fs: FakeFilesystem) -> None:
+    """Reproduce issue from https://github.com/hofbi/dev-tools/issues/113."""
+    root_directory = Path("Test_directory/")
+    fs.create_dir(root_directory)
+    config_file = root_directory / Path(".pre-commit-config.yaml")
+    fs.create_file(config_file)
+    _create_yaml_content(config_file, "foo", extra_attributes={"pass_filenames": 42})
+
+    result = load_hooks(root_directory, config_file)
+
+    assert len(result) == 1
+    assert result[0].id == "check-snake-case"
+
+
+def test_load_hooks_for_repo_builtin_should_parse_config(fs: FakeFilesystem) -> None:
+    """Reproduce issue from https://github.com/hofbi/dev-tools/issues/113."""
+    root_directory = Path("Test_directory/")
+    fs.create_dir(root_directory)
+    config_file = root_directory / Path(".pre-commit-config.yaml")
+    fs.create_file(
+        config_file,
+        contents="""
+repos:
+  - repo: meta
+    hooks:
+      - id: check-hooks-apply
+  - repo: builtin
+    hooks:
+      - id: check-added-large-files
+        exclude: foo
+""",
+    )
+
+    result = load_hooks(root_directory, config_file)
+
+    assert len(result) == 1
+    assert result[0].id == "check-added-large-files"
+
+
+def test_load_hooks_for_minimum_version_attributes_should_parse_config(fs: FakeFilesystem) -> None:
+    """Reproduce issue from https://github.com/hofbi/dev-tools/issues/113."""
+    root_directory = Path("Test_directory/")
+    fs.create_dir(root_directory)
+    config_file = root_directory / Path(".pre-commit-config.yaml")
+    fs.create_file(
+        config_file,
+        contents="""
+minimum_pre_commit_version: 'Update to prek!'
+minimum_prek_version: '0.3.10'
+repos:
+  - repo: meta
+    hooks:
+      - id: check-hooks-apply
+""",
+    )
+
+    result = load_hooks(root_directory, config_file)
+
+    assert len(result) == 0
+
+
+def test_count_excluded_files_for_single_file(fs: FakeFilesystem) -> None:
+    fs.create_file(Path("Repo/test_file.txt"))
+    hook_instance = Hook("test_id", [Path("Repo/test_file.txt")])
+
+    assert hook_instance.count_excluded_files() == 1
+
+
+def test_count_excluded_files_for_multiple_files(fs: FakeFilesystem) -> None:
+    fs.create_file(Path("Repo/file1.txt"))
+    fs.create_file(Path("Repo/file2.txt"))
+    fs.create_file(Path("Repo/file3.txt"))
+    hook_instance = Hook("test_id", [Path("Repo/file1.txt"), Path("Repo/file2.txt"), Path("Repo/file3.txt")])
+
+    assert hook_instance.count_excluded_files() == 3
+
+
+def test_count_excluded_files_for_directory(fs: FakeFilesystem) -> None:
+    fs.create_dir(Path("Repo/test_dir"))
+    fs.create_file(Path("Repo/test_dir/file1.txt"))
+    fs.create_file(Path("Repo/test_dir/file2.txt"))
+    fs.create_file(Path("Repo/test_dir/subdir/file3.txt"))
+    hook_instance = Hook("test_id", [Path("Repo/test_dir")])
+
+    assert hook_instance.count_excluded_files() == 3
+
+
+def test_count_excluded_files_for_mixed_paths(fs: FakeFilesystem) -> None:
+    fs.create_file(Path("Repo/single_file.txt"))
+    fs.create_dir(Path("Repo/test_dir"))
+    fs.create_file(Path("Repo/test_dir/file1.txt"))
+    fs.create_file(Path("Repo/test_dir/file2.txt"))
+    hook_instance = Hook("test_id", [Path("Repo/single_file.txt"), Path("Repo/test_dir")])
+
+    assert hook_instance.count_excluded_files() == 3
+
+
+def test_count_excluded_files_for_non_existing_paths() -> None:
+    hook_instance = Hook("test_id", [Path("Repo/non_existing_file.txt"), Path("Repo/non_existing_dir")])
+
+    assert hook_instance.count_excluded_files() == 0
+
+
+def test_count_excluded_files_for_mixed_existing_and_non_existing_paths(fs: FakeFilesystem) -> None:
+    fs.create_file(Path("Repo/existing_file.txt"))
+    fs.create_dir(Path("Repo/existing_dir"))
+    fs.create_file(Path("Repo/existing_dir/file.txt"))
+    hook_instance = Hook(
+        "test_id",
+        [Path("Repo/existing_file.txt"), Path("Repo/existing_dir"), Path("Repo/non_existing_file.txt")],
+    )
+
+    assert hook_instance.count_excluded_files() == 2
+
+
+def test_count_excluded_files_for_empty_exclude_paths() -> None:
+    hook_instance = Hook("test_id", [])
+
+    assert hook_instance.count_excluded_files() == 0
